requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
